sall_services

+++ src/main/java/it_schoolkg/sall_services/Models/dtos/ProductDTO.java
@@ -11,7 +11,7 @@ import lombok.experimental.FieldDefaults;
 public class ProductDTO {
     Long id;
     String name;
-    long barcode;
+    int barcode;
     boolean active;
 
     CategoriesDTO categories;
