+++ src/main/java/it_schoolkg/sall_services/Models/entities/Product.java
@@ -17,7 +17,7 @@ public class Product {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     Long id;
     String name;
-    long barcode;
+    int barcode;
     boolean active;
     @ManyToOne
     @JoinColumn(name = "id_categories")
